requirements.txt
Flask
python-dotenv
requests
PyMuPDF
anthropic
google-generativeai 